day04/project/requirements.txt
biopython>=1.81
requests>=2.28.0
